+++ src/seb/model_interface.py
@@ -1,10 +1,8 @@
-from typing import Callable, Optional, Protocol, Union, runtime_checkable
+from typing import Any, Callable, Optional, Protocol, Union, runtime_checkable
 
 from numpy import ndarray
 from pydantic import BaseModel
 from torch import Tensor
-
-from .utils import name_to_path
 
 ArrayLike = Union[ndarray, Tensor]
 
@@ -40,12 +38,16 @@ class ModelMeta(BaseModel):
     reference: Optional[str] = None
     languages: list[str] = []
     open_source: bool = False
-    embedding_size: int
+    embedding_size: Optional[int] = None
 
     def get_path_name(self) -> str:
         if self.huggingface_name is None:
-            return name_to_path(self.name)
-        return name_to_path(self.huggingface_name)
+            return self._name_to_path(self.name)
+        return self._name_to_path(self.huggingface_name)
+
+    @staticmethod
+    def _name_to_path(name: str) -> str:
+        return name.replace("/", "__").replace(" ", "_")
 
     def get_huggingface_url(self) -> str:
         if self.huggingface_name is None:
@@ -53,7 +55,12 @@ class ModelMeta(BaseModel):
         return f"https://huggingface.co/{self.huggingface_name}"
 
 
-class SebModel(BaseModel):
+class EmbeddingModel(BaseModel):
+    """
+    An embedding model as implemented in SEB. It notably dynamically loads models (such that models are not loaded when a cache is hit)
+    and includes metadata pertaining to the specific model.
+    """
+
     meta: ModelMeta
     loader: Callable[[], ModelInterface]
     _model: Optional[ModelInterface] = None
@@ -80,7 +87,7 @@ class SebModel(BaseModel):
         self,
         sentences: list[str],
         batch_size: int = 32,
-        **kwargs: dict,
+        **kwargs: Any,
     ) -> ArrayLike:
         """
         Returns a list of embeddings for the given sentences.