+++ src/seb/seb_models/__init__.py
@@ -3,3 +3,4 @@ from .e5_models import *
 from .fairseq_models import *
 from .hf_models import *
 from .openai_models import *
+from .e5_mistral import *